refactor: move MovieDtos to dedicated package
