+++ src/main/java/io/github/janmalch/kino/api/model/movie/MovieDto.java
@@ -1,5 +1,6 @@
-package io.github.janmalch.kino.api.model;
+package io.github.janmalch.kino.api.model.movie;
 
+import io.github.janmalch.kino.api.model.PriceCategoryDto;
 import io.github.janmalch.kino.api.model.presentation.PresentationDto;
 import java.util.List;
