+++ src/main/java/io/github/janmalch/kino/api/model/movie/MovieInfoDto.java
@@ -1,4 +1,4 @@
-package io.github.janmalch.kino.api.model;
+package io.github.janmalch.kino.api.model.movie;
 
 import java.util.Date;
 import java.util.Objects;
